Address token analysis review feedback
