+++ run_neighbor_attention_video.sh
@@ -20,7 +20,7 @@ REPO_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 cd "$REPO_ROOT"
 
 MODEL_DIR="${MODEL_DIR:-$REPO_ROOT/best_models/20260730/best_model}"
-DATADIR="${DATADIR:-/home/yamashita/work_hdd/sample/mini_datasets/j6_2231_fullseq_mini_20260707}"
+DATADIR="${DATADIR:?DATADIR must be set to a dataset directory}"
 PYTHON_BIN="${PYTHON_BIN:-$REPO_ROOT/.venv/bin/python}"
 DEVICE="${DEVICE:-cuda}"
